requirements.txt
# API and HTTP
requests>=2.25.0
python-dotenv>=0.19.0

# Data processing
pandas>=1.3.0
numpy>=1.20.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine learning (optional)
scikit-learn>=1.0.0

# Development tools
pylint>=2.12.0
black>=21.12b0
